Beacon Pressing Improvements
We changed up our beacon approach and button pressing routines. We
finally figured out that we could get much more beacon pressing surface
for end game if we switched our algorith to NOT press the opposing
alliance instead of trying to press ours. Turns out that also simplifies
the code quite a bit, and we can use the same routine in the end-game
where it will automate the final approach and only press the beacon if
the wrong alliance is lit. So it becomes a basic behavior that reduces
the burden on the drive team. We still have more improvements needed -
like correcting our attitude with the imu after the sequence so we can
move on to the next beacon.