Added toggle for switching final DogeCV step (TWO_MINERALS, THREE_MINERALS, ORIGINAL_ALL_MINERALS)
TWO_MINERALS only looks at two minerals, and looks at any gold cubes detected. If a cube is found, MIDDLE or RIGHT is returned based on which side of the screen it is on.
THREE_MINERALS looks at all three minerals. It seems to be more accurate than TWO_MINERALS.
ORIGINAL_ALL_MINERALS looks not only at gold but also silver. It compares relative rather than absolute positition.

Game_6832 now has a toggle for changing this setting, which is only used by DogeCVIntegration. We soon hope to have a more elegant way of doing this.